+++ tests/testthat/test-expr-array.R
@@ -393,3 +393,16 @@ patrick::with_parameters_test_that(
     )
   }
 )
+
+test_that("arr$agg() works", {
+  df <- pl$DataFrame(a = list(c(1, NA), c(42, 13), c(NA, NA)))$cast(pl$Array(pl$Int64, 2))
+
+  expect_equal(
+    df$select(pl$col("a")$arr$agg(pl$element()$null_count())),
+    pl$DataFrame(a = c(1, 0, 2))$cast(pl$UInt32)
+  )
+  expect_equal(
+    df$select(pl$col("a")$arr$agg(pl$element()$drop_nulls())),
+    pl$DataFrame(a = list(1, c(42, 13), numeric(0)))$cast(pl$List(pl$Int64))
+  )
+})